+++ Classes/GeneralPreferencesViewController.h
@@ -6,14 +6,19 @@
 #import "MASPreferencesViewController.h"
 
 @class MainController;
+@class ImageAndTextCell;
 
 @interface GeneralPreferencesViewController : NSViewController <MASPreferencesViewController>{
 @private
 	MainController *mainController;
 	NSDictionary *tableColumns;
+	IBOutlet ImageAndTextCell *imageCell;
 }
 
 - (id)initWithMainController:(MainController*)mainController;
+
+- (IBAction) setStartUrlToCurrent:(id)sender;
+- (void) setStartUrl:(NSURL*)url;
 - (IBAction) open:(id)sender;