+++ Classes/VGTaggerAppDelegate.m
@@ -25,14 +25,13 @@
 @synthesize window, mainController, preferencesWindowController;
 
 - (void)applicationDidFinishLaunching:(NSNotification *)aNotification
-{	
+{	 
 	NSString *path = [[NSBundle mainBundle] pathForResource:@"Vgmdb" ofType:@"rb"];
 	[[MacRuby sharedRuntime] evaluateFileAtPath:path];
 	[[DDTTYLogger sharedInstance ] setLogFormatter:	[[[PSDDFormatter alloc ] init ] autorelease]];
 	[DDLog addLogger:[DDASLLogger sharedInstance]];
 	[DDLog addLogger:[DDTTYLogger sharedInstance]];
 	
-	// Set the fallback values for preferences
 }